moved logging initialization from main
